+++ Bootstrap/Managers/Game.cpp
@@ -20,7 +20,6 @@ char* Game::Developer = NULL;
 char* Game::Name = NULL;
 char* Game::UnityVersion = NULL;
 bool Game::IsIl2Cpp = false;
-bool Game::FirstRun = true;
 
 bool Game::Initialize()
 {
@@ -90,8 +89,7 @@ void Game::ReadAppInfo()
 	std::string appinfopath = std::string(DataPath) + "\\app.info";
 	if (!Core::FileExists(appinfopath.c_str()))
 	{
-		if (FirstRun)
-			Logger::QuickLog("app.info DOES NOT EXIST! Defaulting to UNKNOWN for Company and Product Names", Warning);
+		Logger::QuickLog("app.info DOES NOT EXIST! Defaulting to UNKNOWN for Company and Product Names", Warning);
 		std::string unknown = "UNKNOWN";
 		Developer = new char[unknown.size() + 1];
 		std::copy(unknown.begin(), unknown.end(), Developer);
@@ -117,7 +115,6 @@ void Game::ReadAppInfo()
 			Name[line.size()] = '\0';
 		}
 	appinfofile.close();
-	FirstRun = false;
 	Console::SetDefaultTitleWithGameName();
 }